line_bot/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-language
requests